Fixed: Trakt connection auth tokens not being refreshed
Closes #5285
